pyproject.toml
[tool.poetry]
name = "incarn"
version = "0.1.0"
description = "Bot for discord server, based on python-discord version. Written on Py-Cord."
authors = ["SidVeld Von Disco <sidveld@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.10"
py-cord = "~1.7.3"
GitPython = "~3.1.24"
PyYAML = "~6.0"
aiohttp = "~3.7"
arrow = "~1.2.1"
coloredlogs = "~15.0.1"
deepdiff = "~5.6.0"
rapidfuzz = "^1.8.2"

[tool.poetry.dev-dependencies]
flake8 = "~4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
